refactor(derive): drop dead Espresso calldata authentication path
The calldata data source is only reached pre-ecotone: OpenData routes to the
blob source whenever ecotone is active (ref.Time >= ecotone_time). Every Celo
chain sets ecotone_time = 0, and Espresso (espresso_time) activates strictly
later, so the Espresso event-based batch-authentication branch that #457 folded
into DataFromEVMTransactions could never fire on a real chain.

Restore the calldata data source to its pre-Espresso form. calldata_source.go is
byte-identical to before #457; #457 had split isValidBatchTx into a type/To-only
check plus isAuthorizedBatchSender to serve the blob source's event-based path,
which rippled a needless change into the calldata source. Reinstate the original
5-arg isValidBatchTx (composed from isBatchTxToInbox + isAuthorizedBatchSender)
and point the blob source at isBatchTxToInbox. calldata_source_test.go is also
restored except for one line: DataSourceConfig gained fields (rollupCfg,
batchAuthCaches), so its literal must use named fields instead of positional.

The blob source keeps Espresso event-based authentication and also handles plain
calldata-carrying batcher txs post-ecotone, so no behaviour is lost. Delete the
now-dead Espresso calldata test and move its shared mockAuthEvents helper into
the blob data source test, its only user.

Mirrors the celo-kona removal (drop CeloCalldataSource, use upstream CalldataSource).
