+++ op-node/rollup/derive/calldata_source_test.go
@@ -1,7 +1,6 @@
 package derive
 
 import (
-	"context"
 	"crypto/ecdsa"
 	"math/big"
 	"math/rand"
@@ -116,24 +115,14 @@ func TestDataFromEVMTransactions(t *testing.T) {
 		var expectedData []eth.Data
 		var txs []*types.Transaction
 		for i, tx := range tc.txs {
-			transaction := tx.Create(t, signer, rng)
-			txs = append(txs, transaction)
-
+			txs = append(txs, tx.Create(t, signer, rng))
 			if tx.good {
 				expectedData = append(expectedData, txs[i].Data())
 			}
 		}
 
-		// Legacy mode (no batch authenticator, Espresso inactive) — uses sender-based auth
-		dsCfg := DataSourceConfig{
-			l1Signer:          cfg.L1Signer(),
-			batchInboxAddress: cfg.BatchInboxAddress,
-			rollupCfg:         cfg,
-		}
-		ref := eth.L1BlockRef{Number: 1}
-		// In legacy mode, no L1Fetcher calls are needed for auth (sender check is local)
-		out, err := DataFromEVMTransactions(context.Background(), dsCfg, batcherAddr, txs, nil, ref, testlog.Logger(t, log.LevelCrit))
-		require.NoError(t, err)
+		out := DataFromEVMTransactions(DataSourceConfig{l1Signer: cfg.L1Signer(), batchInboxAddress: cfg.BatchInboxAddress}, batcherAddr, txs, testlog.Logger(t, log.LevelCrit))
 		require.ElementsMatch(t, expectedData, out)
 	}
+
 }
